fix(ci-tests): the integration harness splits at its own seam, and main is green again
#2154 grew internal/compose/integration/harness.go to 505 lines, five past the
500-LOC cap, so scripts/check-go-file-length.sh fails on main's tip and takes
deterministic-gates down with it on every open PR.

The split follows a line the package had already drawn rather than a new one.
seed.go's own opening paragraph names it: raw-SQL seeders there, store-mediated
fixtures on Env in harness.go, "which is the line the two files are split on".
That second half now has its own file. harness.go keeps what a suite needs
before it can act at all — the environment, the pool, the principals — and
harnessfixtures.go holds what it reaches for once it has those.

No behaviour changes; the moved lines are byte-identical bar one. Being moved
makes them new to the strict lint set (new-from-merge-base), which is how a
literal "person" that had sat there unremarked became a goconst finding: it now
uses objPerson and objOrg, the constants the package already had for exactly
this.

seed.go's cross-reference is updated in the same commit, since it named the
file the fixtures just left.
